Save changes
fix an error that was occurring in add action

+++ assets/scripts/expenses/events.js
@@ -39,7 +39,7 @@ const getExpenses = () => {
 
 const onDeleteExpense = (event) => {
   const id = $(event.target).closest('tr').data('id')
-  if (id !== undefined) {
+  if (id !== undefined && id !== '' && id !== 'ID') {
     event.preventDefault()
     api.deleteExpense(id)
       .then(function () {
